requirements.txt
einops==0.3.2
torchprofile
pandas
ujson
future
scikit-learn
scipy==1.10.1
numpy==1.23.5
pytorch-lightning>=1.5
argparse
h5py
ptflops
scikit-image
pywavelets
seaborn
torchinfo
